requirements.txt
pyyaml>=6.0
jsonschema>=4.17.0
